examples/auth-sandbox/requirements.txt
Flask
Flask-Admin
flask-mongoengine
Flask-Login
flask-security
